Ignore live-only tests when in Playback or Record modes
Partially fixes #7416
