+++ sdk/keyvault/Azure.Security.KeyVault.Keys/samples/Azure.Security.KeyVault.Keys.Samples.csproj
@@ -2,6 +2,7 @@
   <PropertyGroup>
     <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
   </PropertyGroup>
+  <Import Project="..\..\..\core\Azure.Core\tests\LiveOnly.props" />
   <ItemGroup>
     <PackageReference Include="nunit" />
     <PackageReference Include="NUnit3TestAdapter" />
