+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/samples/Sample3_GetSecrets.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See License.txt in the project root for
 // license information.
 
+using Azure.Core.Testing;
 using Azure.Identity;
 using NUnit.Framework;
 using System;
@@ -16,7 +17,7 @@ namespace Azure.Security.KeyVault.Secrets.Samples
     /// and list deleted secrets in a soft-delete enabled key vault
     /// using the synchronous methods of the SecretClient.
     /// </summary>
-    [Category("Live")]
+    [LiveOnly]
     public partial class GetSecrets
     {
         [Test]
